Make backups survive app update + add some cleanup
Persistent data files are deleted from the app's sandbox when the
app is updated, unless those files are specifically named at build
time. But since we don't know what our backup files are going to
be called or even how many of them we might need to create, we
can't do that. Legato does not currently have a friendly solution
to this problem that also permits atomic file operations. So we
are forced to (temporarily) run the app unsandboxed.