Support trigger outputs for sensors
Make it possible Input and Output resources that are trigger type
accept non-trigger type samples.  The value is discarded when a
non-trigger sample arrives at a trigger Input or Output resource.

Add the trigger output to periodic sensors implemented using the
periodicSensor component.

Add a fake temperature sensor to the sensor unit test app that
uses the periodicSensor component.