Add query_ReadBufferSampleX() functions
Add the following functions to the Data Hub Query API:
- query_ReadBufferSampleTimestamp()
- query_ReadBufferSampleBoolean()
- query_ReadBufferSampleNumeric()
- query_ReadBufferSampleString()
- query_ReadBufferSampleJson()

Fix crash when reading a buffer of size 0.

Defend against timestamps that go backward so we don't
allow the order of timestamps in a buffer to ever go back
in time, because that could cause infinite loops when
reading the buffer.